user_interaction_service/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-annoying==0.10.6
djangorestframework==3.12.4
psycopg2
psycopg2-binary>=2.8
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
python-dotenv
gunicorn
requests
redis
django-cors-headers
pika
pyyaml
uritemplate
coreapi
